Using HandleBars Template to render common navigation bar code
These changes currently work only Firefox as I was unable to circumvent the same origin policy on Google Chrome. The suggestion that most people offered is to change the path to load the content from a server. We can identify the possible alternatives later once the template rendering code has been written